+++ src/main/java/com/dta/services/dao/ICategoryDao.java
@@ -6,7 +6,7 @@ import com.dta.services.model.Category;
 
 public interface ICategoryDao {
 	void create(Category category);
-	Category getById(int id);
+	Category getById(long id);
 	List<Category> list();
 	void update(Category category);
 	void delete(int id);
