+++ src/main/java/com/dta/services/dao/IMessageDao.java
@@ -6,7 +6,7 @@ import com.dta.services.model.Message;
 
 public interface IMessageDao {
 	void create(Message message);
-	Message getById(int id);
+	Message getById(long id);
 	List<Message> list();
 	void update(Message message);
 	void delete(int id);
